1570 Move Agent, CurrentAgentContext, ToolErrors to platform-ai-agent-api
