--- server/ee/libs/ai/mcp/mcp-tool/mcp-tool-api/build.gradle.kts
@@ -1,8 +1,0 @@
-dependencies {
-    implementation("org.slf4j:slf4j-api")
-    implementation("tools.jackson.core:jackson-databind")
-
-    testImplementation("org.assertj:assertj-core")
-    testImplementation("org.junit.jupiter:junit-jupiter")
-    testImplementation("org.mockito:mockito-core")
-}
