+++ server/ee/libs/platform/platform-ai/platform-ai-agent/platform-ai-agent-api/build.gradle.kts
@@ -1,3 +1,10 @@
 dependencies {
     api("org.springframework.ai:spring-ai-client-chat")
+
+    implementation("org.slf4j:slf4j-api")
+    implementation("tools.jackson.core:jackson-databind")
+
+    testImplementation("org.assertj:assertj-core")
+    testImplementation("org.junit.jupiter:junit-jupiter")
+    testImplementation("org.mockito:mockito-core")
 }
