+++ server/ee/libs/platform/platform-ai/platform-ai-agent/platform-ai-agent-api/src/main/java/com/bytechef/ee/platform/ai/agent/tool/Agent.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the Enterprise License.
  */
 
-package com.bytechef.ee.ai.mcp.tool.usage;
+package com.bytechef.ee.platform.ai.agent.tool;
 
 /**
  * @version ee
