+++ server/ee/libs/platform/platform-ai/platform-ai-agent/platform-ai-agent-api/src/test/java/com/bytechef/ee/platform/ai/agent/tool/ToolErrorsTest.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the Enterprise License.
  */
 
-package com.bytechef.ee.ai.mcp.tool.util;
+package com.bytechef.ee.platform.ai.agent.tool;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
